backend-iowarrior: check input, do not allow to execute 10000
See the error messages in the code.

Also, the exit code of the process is now 1 if something wrong happens
(by convention on UNIX an exit code of 0 means success).

In the main() function, remove the else (and remove one indentation
level for the else body) because there are now early returns.

Note that the riech-o-mat graphical application (written in Python)
already checks the valve states, and disallow to execute "10000" (it
shows a dialog window with an error message when trying to execute that
line). This commit serves as a safety net, and will permit to re-use the
backend with other frontends (e.g. from Matlab).