Remove compiled files
It is not a good practice to put generated files in git.

For example if I modify riech-o-mat-backend-iowarrior.c on Fedora, I
compile it, then I install the software on CentOS, running make will
modify the riech-o-mat-backend-iowarrior executable so it's annoying to
do a git pull afterwards when there are modified files.